Force utf8mb4
To ensure emoji don't throw errors, the database and connections to that database should always be in the utf8mb4 character set. It seems Ubuntu (and probably all of Unix) does this by default but Windows might need a little forcing every now and again.

Also performance improvements when dealing with a directory full of .js files.