Tweetconflicts first pass
Tweetconflicts don't actually work yet, y'all. This is why it's a dev branch. I think this code will not crash, or at least not in a way that breaks anything too badly. But it's still a work in progress. Them's the breaks.

STILL TO DO: Make it work? I just realized that my current approach is likely the worst one so I'll be dismantling this code bit by bit and making something actually good. I don't know what else to say. Somedays you eat the b'ar, and somedays, well, he eats you.