Flask Page for Daily Tweets, Timezone Aware
Going to <FLASK_APP_URL>/day will now let you pick a date with a handy date picker and get results in US/Eastern time. Next step is to prompt for a timezone to use. What a time to be alive.

This is still a far cry from the kind of "powerful" functionality I dreamt of originally. It's, y'know, on the roadmap, as they say in the biz.