Adding a license
I'm adding this license file to open source this project. There are several reasons for this, but I've always believed that this project exists first and foremost because I wanted it to, and whoever else who may want to make use of this project should be able to do so. To that end, it only makes sense that open source it.

But I'm putting in the effort to make it official right now because this is the easiest way for me to keep working on it while at my current employer. So.