SMS column fix
This was not committed before, for reasons unknown to me, but it's needed to better support SMS senders who are using email addresses. These are usually spam and this is probably considered an edge case, but for the sake of completeness I'll allow this. exceptionally long email addresses would still fail, but I don't know how likely this is.